requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
django-environ
django-filter
gunicorn
requests
